requirements.txt
berserk==0.13.2
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ndjson==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
